requirements.txt
flask
twython
pandas
sqlalchemy
textblob
flask
flask_restful
kafka-python
